moved classes into own folder, and made api call for arrival time
