+++ src/app/classes/details.ts
@@ -1,0 +1,14 @@
+export class Details {
+    Actual: boolean;
+    BlockNumber: number;
+    DepartureText: string;
+    DepartureTime: string;
+    Description: string;
+    Gate: string;
+    Route: string;
+    RouteDirection: string;
+    Terminal: string;
+    VehicleHeading: number;
+    VehicleLatitude: number;
+    VehicleLongitude: number;
+}
