+++ src/app/classes/stops.ts
