+++ src/app/metroapi.service.ts
@@ -2,9 +2,10 @@ import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs'
 
-import { Route } from './route'
-import { Directions } from './directions';
-import { Stops } from './stops';
+import { Route } from './classes/route'
+import { Directions } from './classes/directions';
+import { Stops } from './classes/stops';
+import { Details } from './classes/details';
 
 @Injectable({
   providedIn: 'root'
@@ -23,5 +24,9 @@ export class MetroApiService {
 
   getStops(direction: any, route: any): Observable<Stops[]> {
     return this.http.get<Stops[]>('http://svc.metrotransit.org/NexTrip/Stops/' + route + '/' + direction)
+  }
+
+  getArrivalTime(route, direction, stop): Observable<Details[]> {
+    return this.http.get<Details[]>('http://svc.metrotransit.org/NexTrip/' + route + '/' + direction + '/' + stop)
   }
 }
